Fix YOLOv5 bbox conversion
