+++ foxai/visualizer.py
@@ -1,8 +1,10 @@
-from typing import Tuple
+from typing import List, Tuple
 
+import cv2
 import matplotlib
 import numpy as np
 import torch
+import torchvision
 from matplotlib.pyplot import Figure
 
 from foxai.array_utils import (
@@ -13,6 +15,7 @@ from foxai.array_utils import (
     standardize_array,
     transpose_array,
 )
+from foxai.explainer.computer_vision.object_detection.types import ObjectDetectionOutput
 
 
 def generate_figure(
@@ -191,3 +194,133 @@ def single_channel_visualization(
         figsize=figsize,
         alpha=alpha,
     )
+
+
+def get_heatmap_bbox(
+    heatmap: np.ndarray,
+    bbox: List[int],
+    mask_value: int = 0,
+) -> np.ndarray:
+    """_summary_
+
+    Code based on https://github.com/pooya-mohammadi/deep_utils/blob/main/deep_utils/utils/box_utils/boxes.py.
+
+    Args:
+        heatmap: Heatmap to visualize.
+        bbox: Bounding box of detection.
+        mask_value: Masking value . Defaults to 0.
+
+    Returns:
+        Numpy array with heatmap only present in area of given bounding box.
+    """
+    # fill the outer area of the selected box
+    mask = np.ones_like(heatmap, dtype=np.uint8) * mask_value
+    mask[bbox[0] : bbox[2], bbox[1] : bbox[3]] = 1
+    masked_heatmap = cv2.multiply(heatmap, mask)
+    return masked_heatmap
+
+
+def draw_heatmap_in_bbox(
+    bbox: List[int],
+    heatmap: torch.Tensor,
+    img: np.ndarray,
+) -> np.ndarray:
+    """Draw heatmap in bounding box on image.
+
+    Args:
+        bbox: List of coordinates for bounding box.
+        heatmap: Heatmap to display.
+        img: Original image.
+
+    Returns:
+        Image with displayed heatmap in bounding box area.
+    """
+    heatmap_np = (
+        heatmap.squeeze(0)
+        .mul(255)
+        .add_(0.5)
+        .clamp_(0, 255)
+        .permute(1, 2, 0)
+        .detach()
+        .cpu()
+        .numpy()
+        .astype(np.uint8)
+    )
+    heatmap_np = cv2.applyColorMap(heatmap_np, cv2.COLORMAP_JET)
+
+    masked_heatmap = get_heatmap_bbox(heatmap=heatmap_np, bbox=bbox).astype(np.float32)
+
+    img = cv2.add(img, masked_heatmap)
+    img = img / img.max()
+    img = (img * 255).astype(np.uint8)
+    return img
+
+
+def concat_images(images: List[np.ndarray]) -> np.ndarray:
+    """Concatenate images into one.
+
+    Args:
+        images: List of images to merge.
+
+    Returns:
+        Final image.
+    """
+    w, h = images[0].shape[:2]
+    width = w
+    height = h * len(images)
+    base_img = np.zeros((width, height, 3), dtype=np.uint8)
+    for i, img in enumerate(images):
+        base_img[:, h * i : h * (i + 1), ...] = img
+
+    return base_img
+
+
+def object_detection_visualization(
+    detections: ObjectDetectionOutput,
+    input_image: torch.Tensor,
+) -> np.ndarray:
+    """Create array with detection heatmaps.
+
+    Args:
+        detections: Object detection data class.
+        input_image: Image in shape (C x H x W) or (H x W).
+
+    Returns:
+        Array of series of images with heatmap displayed on detection bounding boxes.
+    """
+    masks = detections.saliency_maps
+    boxes = [pred.bbox for pred in detections.predictions]
+    class_names = [pred.class_name for pred in detections.predictions]
+    img_to_display = (
+        input_image.squeeze(0)
+        .mul(255)
+        .add_(0.5)
+        .clamp_(0, 255)
+        .permute(1, 2, 0)
+        .detach()
+        .cpu()
+        .numpy()
+    )
+    img_to_display = img_to_display[..., ::-1]  # convert to bgr
+    images = [img_to_display]
+
+    for i, mask in enumerate(masks):
+        res_img = img_to_display.copy()
+        bbox, cls_name = boxes[i], class_names[i]
+        bbox = [int(val) for val in bbox]
+        res_img = draw_heatmap_in_bbox(bbox, mask, res_img)
+
+        # convert to (C x H x W)
+        res_img_tensor = torch.tensor(res_img).transpose(0, 2).transpose(1, 2)
+        bbox = [bbox[1], bbox[0], bbox[3], bbox[2]]
+        res_img_tensor = torchvision.utils.draw_bounding_boxes(
+            image=res_img_tensor,
+            boxes=torch.tensor([bbox]),
+            labels=[cls_name],
+        )
+        # convert to (H x W x C)
+        res_img = res_img_tensor.transpose(1, 2).transpose(0, 2).numpy()
+        images.append(res_img)
+
+    final_image = concat_images(images)
+    return final_image
